+++ CHANGELOG.md
@@ -4,7 +4,12 @@ All notable changes to the ASK CLI Configure extension will be documented in thi
 ## [Unreleased]
 - Looking for new requirements
 
+## 0.2.0 - (2018-08-12)
+### Added
+- Command: Copy ASK profile name from 'cli_config'...
+
+
 ## 0.1.0 - (2018-08-11)
 ### Added
-- Command: Open config file
+- Command: Open `.ask/cli_config` file
 - Command: Browse online docs